move useChat to channel folder, error handling in component
