+++ src/hooks/channel/useChat.js
@@ -1,7 +1,6 @@
 import useSWR from 'swr';
 
-import ErrorBox from '../components/ErrorBox';
-import { API } from '../constants/api';
+import { API } from '../../constants/api';
 
 async function fetcher(channelId) {
   try {
@@ -14,15 +13,14 @@ async function fetcher(channelId) {
     const { result, data, message } = await response.json();
 
     if (result === 'error') {
-      return <ErrorBox message={message} />;
+      throw new Error(message);
     }
 
     return data;
   } catch (err) {
-    return <ErrorBox message={err.message} />;
+    return err;
   }
 }
-
 export default function useChat(channelId) {
   const { data: chatList, mutate } = useSWR(
     channelId ? '/chat/channelId' : null,
